refactor(daemon): extract prompt text helpers
